Moved createRangeEntity and createMeleeEntity from WeaponConfig to WeaponFactory
